+++ bin/xdmod-ingestor
@@ -77,7 +77,7 @@ function main()
         exit(1);
     }
 
-    $help = $ingest = $aggregate = $aggregateCloud = $noAppend = $ingestAll = $ingestShredded
+    $help = $ingest = $aggregate = $noAppend = $ingestAll = $ingestShredded
         = $ingestStaging = $ingestHpcdb = $datatype = false;
 
     $startDate = $endDate = $lastModifiedStartDate = $realmToAggregate = $datatypeValue = null;
@@ -266,8 +266,8 @@ function main()
                     $dwi->ingestCloudDataOpenStack();
                 }
 
-                if($datatypeValue == 'generalcloud'){
-                    $dwi->ingestCloudDataGeneral();
+                if($datatypeValue == 'genericcloud'){
+                    $dwi->ingestCloudDataGeneric();
                 }
             }
         } catch (Exception $e) {
@@ -284,7 +284,7 @@ function main()
         $logger->info('Aggregating data');
         try {
             //If there is no realm specified to aggregate then all realms should be aggregated
-            if($realmToAggregate === null){
+            if($realmToAggregate == 'job' || $realmToAggregate === null){
                 $dwi->aggregateAllJobs($lastModifiedStartDate);
             }
