+++ bin/xdmod-shredder
@@ -233,7 +233,7 @@ function main()
         $logger->notice("Job errors written to '$jobErrorLogFile'");
     }
 
-    if (!$dryRun && !in_array($format, array('openstack', 'generalcloud'))) {
+    if (!$dryRun && !in_array($format, array('openstack', 'genericcloud'))) {
         $logger->notice('Normalizing data');
 
         try {
