+++ classes/OpenXdmod/Shredder/Cloud.php
@@ -1,0 +1,61 @@
+<?php
+/**
+ * Abstract class used when shredding cloud event data.
+ *
+ * @author Greg Dean <gmdean@ccr.buffalo.edu>
+ */
+namespace OpenXdmod\Shredder;
+
+use Exception;
+use DateTime;
+use DateTimeZone;
+use CCR\DB\iDatabase;
+use OpenXdmod\Shredder;
+use ETL\Utilities;
+
+abstract class Cloud extends Shredder{
+
+  /**
+   * @inheritdoc
+   */
+  public function __construct(iDatabase $db)
+  {
+      $this->logger = \Log::singleton('null');
+  }
+
+  /**
+   * Shredding by specifing a single file in not supported by the cloud pipelines.
+   * Throw an exception if someone tries to shred cloud data using the -i flag instead
+   * of using -d
+   */
+  public function shredFile($line)
+  {
+      throw new Exception('Cloud resources do not support shredding by file. Please use the -d option and specify a directory');
+  }
+
+  /**
+   * @inheritdoc
+   */
+  public function shredDirectory($directory, array $pipelines)
+  {
+      if (!is_dir($directory)) {
+          $this->logger->err("'$directory' is not a directory");
+          return false;
+      }
+
+      if (empty($pipelines)) {
+          $this->logger->err("A pipeline to run was not specified. Please provide a pipeline to run.");
+          return false;
+      }
+
+      Utilities::runEtlPipeline(array('jobs-common','jobs-cloud-common','ingest-resources'), $this->logger);
+      Utilities::runEtlPipeline(
+          $pipelines,
+          $this->logger,
+          array(
+            'include-only-resource-codes' => $this->resource,
+            'variable-overrides' => ['CLOUD_EVENT_LOG_DIRECTORY' => $directory]
+          )
+      );
+  }
+}
