+++ open_xdmod/modules/xdmod/integration_tests/scripts/bootstrap.sh
@@ -30,6 +30,7 @@ then
         sudo -u xdmod xdmod-shredder -r `basename $resource .log` -f slurm -i $resource;
     done
     sudo -u xdmod xdmod-ingestor
+    sudo -u xdmod xdmod-shredder -r openstack -d $REF_DIR/openstack -f openstack
     sudo -u xdmod xdmod-import-csv -t names -i $REF_DIR/names.csv
     sudo -u xdmod xdmod-ingestor
     php /root/bin/createusers.php
@@ -38,13 +39,6 @@ then
     sudo -u xdmod php /usr/share/xdmod/tools/etl/etl_overseer.php -p xdmod.acls-import
     #Updating minmaxdate table so data for cloud realm shows up
     mysql -e "UPDATE modw.minmaxdate SET max_job_date = '2018-07-01';"
-    #Ingesting cloud data from references folder
-    sudo -u xdmod php /usr/share/xdmod/tools/etl/etl_overseer.php -p jobs-common
-    sudo -u xdmod php /usr/share/xdmod/tools/etl/etl_overseer.php -p jobs-cloud-common
-    sudo -u xdmod php /usr/share/xdmod/tools/etl/etl_overseer.php -p ingest-resources
-    sudo -u xdmod php /usr/share/xdmod/tools/etl/etl_overseer.php -p jobs-cloud-ingest-openstack -r openstack -d "CLOUD_EVENT_LOG_DIRECTORY=$REF_DIR/openstack"
-    sudo -u xdmod php /usr/share/xdmod/tools/etl/etl_overseer.php -p jobs-cloud-extract-openstack
-    sudo -u xdmod php /usr/share/xdmod/tools/etl/etl_overseer.php -p cloud-state-pipeline
 fi
 
 if [ "$XDMOD_TEST_MODE" = "upgrade" ];
